refactor factory function
